Declare struct sim_dirent the same way as rockbox core's  struct dirent. Fixes file time reading for 64 bit sims.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@13053 a1c6a512-1295-4272-9138-f99709370657

+++ uisimulator/common/io.c
@@ -67,8 +67,8 @@ void dircache_rename(const char *oldpath, const char *newpath);
 struct sim_dirent {
     unsigned char d_name[MAX_PATH];
     int attribute;
-    int size;
-    int startcluster;
+    long size;
+    long startcluster;
     unsigned short wrtdate; /*  Last write date */ 
     unsigned short wrttime; /*  Last write time */
 };
